src/requirements.txt
aiobotocore==2.13.1
aiohappyeyeballs==2.3.5
aiohttp==3.10.2
aioitertools==0.11.0
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
anyconfig==0.14.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==24.2.0
Babel==2.14.0
bayesian-optimization==1.5.1
beautifulsoup4==4.12.3
binaryornot==0.4.4
bleach==6.1.0
boto3==1.35.13
botocore==1.35.13
Brotli==1.1.0
build==1.2.1
cached-property==1.5.2
cachetools==5.4.0
certifi==2024.7.4
cffi==1.17.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cookiecutter==2.6.0
cycler==0.12.1
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
dynaconf==3.2.6
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.2.2
executing==2.0.1
fastjsonschema==2.20.0
feature-engine==1.8.0
fonttools==4.53.1
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.6.0
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
hyperframe==6.0.1
idna==3.7
imbalanced-learn==0.12.0
importlib-metadata==6.10.0
importlib_resources==6.4.0
ipykernel==6.29.0
ipython==8.18.1
ipywidgets==8.1.3
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter_client==8.6.2
jupyter-console==6.6.3
jupyter_core==5.7.2
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.11
kedro==0.18.14
kedro-datasets==2.0.0
kiwisolver==1.4.5
lazy_loader==0.4
llvmlite==0.43.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
more-itertools==10.4.0
multidict==6.0.5
munkres==1.1.4
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest_asyncio==1.6.0
notebook==7.2.1
notebook_shim==0.2.4
numba==0.60.0
numpy==1.26.0
omegaconf==2.3.0
openpyxl==3.1.5
overrides==7.7.0
packaging==24.1
pandas==2.2.0
pandocfilters==1.5.0
parse==1.20.2
parso==0.8.4
patsy==0.5.6
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pip==24.2
pip-tools==7.4.1
pkgutil_resolve_name==1.3.10
platformdirs==4.2.2
pluggy==1.2.0
polars==0.20.25
prometheus_client==0.20.0
prompt_toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyaml==24.9.0
pyarrow==16.0.0
pycparser==2.22
Pygments==2.18.0
pyngrok==7.2.0
pyparsing==3.1.2
pyproject_hooks==1.1.0
PyQt5==5.15.11
#PyQt5-Qt5==5.15.11
PyQt5_sip==12.15.0
PySocks==1.7.1
python-dateutil==2.9.0
python-json-logger==2.0.7
python-slugify==8.0.4
pytoolconfig==1.2.5
pytz==2024.1
#PyYAML==6.0
pyzmq==26.1.0
qtconsole==5.5.2
QtPy==2.4.1
referencing==0.35.1
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rope==1.13.0
rpds-py==0.20.0
s3fs==2024.6.0
s3transfer==0.10.2
scikit-learn==1.4.0
scikit-optimize==0.10.2
scikit-posthocs==0.9.0
scipy==1.13.1
seaborn==0.13.2
Send2Trash==1.8.3
setuptools==72.1.0
shap==0.46.0
six==1.16.0
slicer==0.0.8
smmap==5.0.0
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.2
statsmodels==0.14.2
terminado==0.18.1
text-unidecode==1.3
threadpoolctl==3.5.0
tinycss2==1.3.0
toml==0.10.2
tomli==2.0.1
toposort==1.10
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
typing-utils==0.1.0
tzdata==2024.1
unicodedata2==15.1.0
uri-template==1.3.0
urllib3==1.26.19
wcwidth==0.2.13
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
wheel==0.44.0
widgetsnbextension==4.0.11
wrapt==1.16.0
xgboost==2.0.3
yarl==1.9.4
zipp==3.19.2